Move schema test to proper place
